# docs/api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A multiplayer FPS game server with OAuth authentication and N:1 account linking",
        "title": "LIFE Game Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/homveloper/life",
            "email": "support@lifegame.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/homveloper/life/blob/main/LICENSE"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/api/v1/auth.GuestLogin": {
            "post": {
                "description": "Login as guest user using device identifier for immediate game access",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Guest login with device ID",
                "parameters": [
                    {
                        "description": "JSON-RPC request with GuestLoginRequest params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_GuestLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token for guest user",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_GuestLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth.LinkSocial": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Convert guest account to social account by linking with OAuth provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Link guest account to social provider",
                "parameters": [
                    {
                        "description": "JSON-RPC request with LinkSocialRequest params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_LinkSocialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated JWT token with social account",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_LinkSocialResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not a guest account or already linked",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth.OAuthCallback": {
            "post": {
                "description": "Complete OAuth authentication with authorization code and receive JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Complete OAuth authentication flow",
                "parameters": [
                    {
                        "description": "JSON-RPC request with OAuthCallbackRequest params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_OAuthCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token and user information",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_OAuthCallbackResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters or authorization code",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth.OAuthStart": {
            "post": {
                "description": "Initiate OAuth authentication with a supported provider (google, github, discord)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Start OAuth authentication flow",
                "parameters": [
                    {
                        "description": "JSON-RPC request with OAuthStartRequest params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_OAuthStartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OAuth authorization URL",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_OAuthStartResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                }
            }
        },
...
        "github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_GuestLoginRequest": {
            "type": "object",
            "properties": {
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "method": {
                    "type": "string"
                },
                "params": {
                    "$ref": "#/definitions/internal_api_handlers.GuestLoginRequest"
                }
            }
        },
        "github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_LinkSocialRequest": {
            "type": "object",
            "properties": {
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "method": {
                    "type": "string"
                },
                "params": {
                    "$ref": "#/definitions/internal_api_handlers.LinkSocialRequest"
                }
            }
        },
...
        "github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_OAuthCallbackRequest": {
            "type": "object",
            "properties": {
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "method": {
                    "type": "string"
                },
                "params": {
                    "$ref": "#/definitions/internal_api_handlers.OAuthCallbackRequest"
                }
            }
        },
        "github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_OAuthStartRequest": {
            "type": "object",
            "properties": {
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "method": {
                    "type": "string"
                },
                "params": {
                    "$ref": "#/definitions/internal_api_handlers.OAuthStartRequest"
                }
            }
        },
...
        "github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_GuestLoginResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "$ref": "#/definitions/internal_api_handlers.GuestLoginResponse"
                }
            }
        },
        "github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_LinkSocialResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "$ref": "#/definitions/internal_api_handlers.LinkSocialResponse"
                }
            }
        },
...
        "github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_OAuthCallbackResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "$ref": "#/definitions/internal_api_handlers.OAuthCallbackResponse"
                }
            }
        },
        "github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_OAuthStartResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "$ref": "#/definitions/internal_api_handlers.OAuthStartResponse"
                }
            }
        },
...
        "internal_api_handlers.GuestLoginRequest": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}